obstacles and more

+++ Assets/Scripts/Game/Run/SeedEncoder.cs
@@ -8,7 +8,7 @@ namespace Game.Run
         // translation table for values 0 to 15
         static readonly char[] CharFromCode = { 'A', 'C', 'D', 'E', 'F', 'I', 'L', 'M', 'N', 'O', 'P', 'R', 'S', 'T', 'U', 'Y' };
         // translation table for chars 'A' to 'Z'
-        static readonly ulong[] CodeFromChar = { 0, 16, 1, 2, 3, 4, 17, 18, 5, 19, 20, 6, 7, 8, 9, 10, 21, 11, 12, 13, 14, 22, 23, 24, 15, 25 };
+        static readonly uint[] CodeFromChar = { 0, 16, 1, 2, 3, 4, 17, 18, 5, 19, 20, 6, 7, 8, 9, 10, 21, 11, 12, 13, 14, 22, 23, 24, 15, 25 };
         static readonly ulong BaseSeed = 0xBA5E_5EED_0123_1006;
         static readonly int RotationStep = 36;
 
@@ -41,7 +41,7 @@ namespace Game.Run
             return seed;
         }
 
-        static ulong DecodeChar(char c) => c is >= 'A' and <= 'Z' ? CodeFromChar[c - 'A'] : c;
+        static uint DecodeChar(char c) => c is >= 'A' and <= 'Z' ? CodeFromChar[c - 'A'] : c;
 
         static string GetRandomSeed()
         {
